test: split sourceless DAGNode contract into its own file
The dynamic single-step flow helper and its test in
test_graph_structure.py were testing the FunctionSpec hook contract
(sourceless DAGNode construction via _create_sourceless_single_step_node),
not graph topology. Move them into test_sourceless_dag_node.py with a
docstring that names the contract directly so a future contributor
does not have to follow the chain helper -> graph.py module comment ->
FunctionSpec backstory.

No behavior change.
